Separate test utility classes and split them up
