+++ Assets/Tests/EditMode/Infrastructure/Services/Scene/SceneStrategyProviderTests.cs
@@ -3,7 +3,7 @@ using NUnit.Framework;
 using WizardSpells.Architecture.GameStates.InitialSettingStrategies;
 using WizardSpells.Data.Static.Enumerations.Scene;
 using WizardSpells.Infrastructure.Services.Scene;
-using WizardSpells.Tests.EditMode.TestUtilities;
+using WizardSpells.Tests.Utilities.Infrastructure.Services.Scene;
 using WizardSpells.Utilities.Patterns.Strategy;
 
 namespace WizardSpells.Tests.EditMode.Infrastructure.Services.Scene
