--- Assets/Tests/EditMode/TestUtilities.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 542c898af66a48d6b15f62453cf422de
-timeCreated: 1674728560
