--- Assets/Tests/EditMode/TestUtilities/Create.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 0ee4051b8dd74d27a96e6f1f4bdabb7e
-timeCreated: 1672962667
