--- Assets/Tests/EditMode/TestUtilities/Substitute.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 8ccc718530ac4407ad071c2c37e4af94
-timeCreated: 1674718874
