--- Assets/Tests/PlayMode/TestUtilities.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: cf8d9f44b114475aa4d408079d4eeb59
-timeCreated: 1674729098
