--- Assets/Tests/PlayMode/TestUtilities/Create.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4462b41e180daac46987ed428f2607fd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
