--- Assets/Tests/PlayMode/TestUtilities/Setup.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: ef3347508b2745e8bafed7bec7bc8612
-timeCreated: 1673104512
