+++ Assets/Tests/PlayMode/WizardSpells.Tests.PlayMode.asmdef
@@ -4,6 +4,7 @@
     "references": [
         "GUID:27619889b8ba8c24980f49ee34dbb44a",
         "GUID:0acc523941302664db1f4e527237feb3",
+        "GUID:0680a52685fd4734ab54d88dddbe1068",
         "GUID:3b29a6033bac96c42800d74155872b33"
     ],
     "includePlatforms": [],
