+++ Assets/Tests/Utilities/Common/Extensions/CreateExtensions.cs
@@ -1,0 +1,7 @@
+﻿namespace WizardSpells.Tests.Utilities.Common.Extensions
+{
+    public static class CreateExtensions
+    {
+        public static T NewIfDefault<T>(this T obj) where T : class, new() => obj ?? new T();
+    }
+}
