+++ Assets/Tests/Utilities/Common/Extensions/SubstituteExtensions.cs
@@ -1,0 +1,9 @@
+﻿using NSubstitute;
+
+namespace WizardSpells.Tests.Utilities.Common.Extensions
+{
+    public static class SubstituteExtensions
+    {
+        public static T SubstituteIfDefault<T>(this T obj) where T : class => obj ?? Substitute.For<T>();
+    }
+}
