+++ Assets/Tests/Utilities/Features/Services/Force/MotionForce/Create.cs
@@ -1,0 +1,9 @@
+﻿using WizardSpells.Features.Services.Force.MotionForce;
+
+namespace WizardSpells.Tests.Utilities.Features.Services.Force.MotionForce
+{
+    public static class Create
+    {
+        public static MotionForceAccumulator MotionForceAccumulator() => new();
+    }
+}
