+++ Assets/Tests/Utilities/Infrastructure/Factories/Components/TestObject.cs
@@ -1,0 +1,10 @@
+﻿using UnityEngine;
+using Zenject;
+
+namespace WizardSpells.Tests.Utilities.Infrastructure.Factories.Components
+{
+    public class TestObject : MonoBehaviour
+    {
+        [Inject] public float FloatValue { get; }
+    }
+}
