+++ Assets/Tests/Utilities/WizardSpells.Tests.Utilities.asmdef
@@ -1,0 +1,20 @@
+{
+  "name": "WizardSpells.Tests.Utilities",
+  "rootNamespace": "WizardSpells",
+  "references": [
+    "GUID:3b29a6033bac96c42800d74155872b33",
+    "GUID:0d8beb7f090555447a6cf5ce9e54dbb4"
+  ],
+  "includePlatforms": [],
+  "excludePlatforms": [],
+  "allowUnsafeCode": false,
+  "overrideReferences": true,
+  "precompiledReferences": [
+    "NSubstitute.dll",
+    "Zenject-usage.dll"
+  ],
+  "autoReferenced": false,
+  "defineConstraints": [],
+  "versionDefines": [],
+  "noEngineReferences": false
+}
