+++ Assets/Tests/Utilities/WizardSpells.Tests.Utilities.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0680a52685fd4734ab54d88dddbe1068
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
